requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.10.3
django-extensions==1.7.4
django-jsonview==1.0.0
djangorestframework==3.5.3
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.12.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
